Added: Interaction table modifiers r and m. BugFix: Chromatin node radius is interpreted as double (used to be int)

+++ source/Chromatin/Chromatin_config.cpp
@@ -51,7 +51,7 @@ void Chromatin::consistancy_check(){
             errorMessage+= TRESET;
             throw std::runtime_error(errorMessage);
         }
-        Node_radius=stoi(Node_radius_stat);
+        Node_radius=stod(Node_radius_stat);
     }
     
     if (BondNominalLength_stat!= "Au" && BondNominalLength_stat!= "Av") {
